Create Baggage API to match specification
