+++ bridge/opentracing/internal/mock.go
@@ -23,9 +23,9 @@ import (
 
 	oteltrace "go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/internal/baggage"
 	otelparent "go.opentelemetry.io/otel/internal/trace/parent"
 	"go.opentelemetry.io/otel/label"
-	otelpropagators "go.opentelemetry.io/otel/propagators"
 
 	"go.opentelemetry.io/otel/bridge/opentracing/migration"
 )
@@ -45,7 +45,7 @@ type MockContextKeyValue struct {
 }
 
 type MockTracer struct {
-	Resources             otelpropagators.Map
+	Resources             baggage.Map
 	FinishedSpans         []*MockSpan
 	SpareTraceIDs         []oteltrace.ID
 	SpareSpanIDs          []oteltrace.SpanID
@@ -60,7 +60,7 @@ var _ migration.DeferredContextSetupTracerExtension = &MockTracer{}
 
 func NewMockTracer() *MockTracer {
 	return &MockTracer{
-		Resources:             otelpropagators.NewEmptyMap(),
+		Resources:             baggage.NewEmptyMap(),
 		FinishedSpans:         nil,
 		SpareTraceIDs:         nil,
 		SpareSpanIDs:          nil,
@@ -86,7 +86,7 @@ func (t *MockTracer) Start(ctx context.Context, name string, opts ...oteltrace.S
 		officialTracer: t,
 		spanContext:    spanContext,
 		recording:      config.Record,
-		Attributes: otelpropagators.NewMap(otelpropagators.MapUpdate{
+		Attributes: baggage.NewMap(baggage.MapUpdate{
 			MultiKV: config.Attributes,
 		}),
 		StartTime:    startTime,
@@ -179,10 +179,10 @@ func (t *MockTracer) DeferredContextSetupHook(ctx context.Context, span oteltrac
 }
 
 type MockEvent struct {
-	CtxAttributes otelpropagators.Map
+	CtxAttributes baggage.Map
 	Timestamp     time.Time
 	Name          string
-	Attributes    otelpropagators.Map
+	Attributes    baggage.Map
 }
 
 type MockSpan struct {
@@ -192,7 +192,7 @@ type MockSpan struct {
 	SpanKind       oteltrace.SpanKind
 	recording      bool
 
-	Attributes   otelpropagators.Map
+	Attributes   baggage.Map
 	StartTime    time.Time
 	EndTime      time.Time
 	ParentSpanID oteltrace.SpanID
@@ -223,7 +223,7 @@ func (s *MockSpan) SetError(v bool) {
 }
 
 func (s *MockSpan) SetAttributes(attributes ...label.KeyValue) {
-	s.applyUpdate(otelpropagators.MapUpdate{
+	s.applyUpdate(baggage.MapUpdate{
 		MultiKV: attributes,
 	})
 }
@@ -232,7 +232,7 @@ func (s *MockSpan) SetAttribute(k string, v interface{}) {
 	s.SetAttributes(label.Any(k, v))
 }
 
-func (s *MockSpan) applyUpdate(update otelpropagators.MapUpdate) {
+func (s *MockSpan) applyUpdate(update baggage.MapUpdate) {
 	s.Attributes = s.Attributes.Apply(update)
 }
 
@@ -287,10 +287,10 @@ func (s *MockSpan) AddEvent(ctx context.Context, name string, attrs ...label.Key
 
 func (s *MockSpan) AddEventWithTimestamp(ctx context.Context, timestamp time.Time, name string, attrs ...label.KeyValue) {
 	s.Events = append(s.Events, MockEvent{
-		CtxAttributes: otelpropagators.MapFromContext(ctx),
+		CtxAttributes: baggage.MapFromContext(ctx),
 		Timestamp:     timestamp,
 		Name:          name,
-		Attributes: otelpropagators.NewMap(otelpropagators.MapUpdate{
+		Attributes: baggage.NewMap(baggage.MapUpdate{
 			MultiKV: attrs,
 		}),
 	})
